moved units to model
