+++ model/co2calcs.py
@@ -18,8 +18,7 @@ import model.fairutil
 
 from model.data_handler import DataHandler
 from model.decorators import data_func
-
-from tools.units import map_to_unit, Mt
+from model.units import map_to_unit, Mt
 
 C_TO_CO2EQ = 3.666
 # Note: a different value of 3.64 is sometimes used for certain results in Excel
